+++ internal/docs/guides.js
@@ -1,13 +1,14 @@
-import GuidingPrinciples from './pages/guiding-principles'
-import Usage from './pages/usage'
-import ContributionGuide from './pages/contribution-guide'
-import FAQS from './pages/faqs'
-import Changes from './pages/changes'
-import AutomationGlossary from './pages/automation-glossary'
-import MigrationGuide from './pages/migration-guide'
+import GuidingPrinciples from './guides/guiding-principles'
+import Usage from './guides/usage'
+import ContributionGuide from './guides/contribution-guide'
+import FAQS from './guides/faqs'
+import Changes from './guides/changes'
+import AutomationGlossary from './guides/automation-glossary'
+import MigrationGuide from './guides/migration-guide'
+import GettingStarted from './guides/getting-started'
 
 const guides = [
-  { path: '/', title: 'Getting started' },
+  { path: '/', title: 'Getting started', component: GettingStarted },
   { path: '/usage', title: 'How to use Cosmos?', component: Usage },
   { path: '/guiding-principles', title: 'Guiding Principles', component: GuidingPrinciples },
   { path: '/faqs', title: 'FAQs', component: FAQS },
