+++ internal/docs/guides/automation-glossary.js
