+++ internal/docs/guides/contribution-guide/architecture.js
