+++ internal/docs/guides/contribution-guide/component-files.js
